requirements/base.txt
Django==4.2
djangorestframework==3.14.0
Pillow==9.5.0
django-jazzmin==2.6.0
python-dotenv==1.0.0
drf-yasg==1.21.5
psycopg2-binary==2.9.6
phonenumbers==8.13.9
django-phonenumber-field==7.0.2
rest-framework-simplejwt==0.0.2
django-recaptcha==3.0.0
environs[django]==9.5.0
django-modeltranslation==0.18.9
django-rosetta==0.9.9



